services/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
structlog==23.2.0
prometheus-client==0.19.0
redis==5.0.1
aioredis==2.0.1
PyJWT==2.8.0
bcrypt==4.0.1
slowapi==0.1.9
limits==3.7.0
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
psutil==5.9.8
email-validator==2.1.0.post1